detele Test_All_HomeViewModel and update Test_Home_ViewModel
